+++ packages/client/src/tx/submit-fns.ts
@@ -1,60 +1,123 @@
-import { HexString } from "@polkadot-api/substrate-bindings"
+import { Blake2256, HexString } from "@polkadot-api/substrate-bindings"
 import {
   EMPTY,
   Observable,
-  combineLatest,
   concat,
+  distinctUntilChanged,
   filter,
   lastValueFrom,
   map,
   mergeMap,
   of,
   take,
-  takeWhile,
 } from "rxjs"
-import { ChainHead$, SystemEvent } from "@polkadot-api/observable-client"
+import {
+  ChainHead$,
+  PinnedBlocks,
+  SystemEvent,
+} from "@polkadot-api/observable-client"
 import { AnalyzedBlock } from "@polkadot-api/observable-client"
-import { TxBroadcastEvent } from "./types"
-import { TxEventsPayload } from "./types"
+import { TxEvent, TxEventsPayload, TxFinalizedPayload } from "./types"
+import { continueWith } from "@/utils"
+import { fromHex, toHex } from "@polkadot-api/utils"
 
-const tx$ = (
-  chainHead: ChainHead$,
-  broadcastTx$: (tx: string) => Observable<never>,
-  tx: string,
-  _at?: HexString,
+// TODO: make it dynamic based on the tx-function of the client
+const hashFromTx = (tx: HexString) => toHex(Blake2256(fromHex(tx)))
+
+const computeState = (
+  analized$: Observable<AnalyzedBlock>,
+  blocks$: Observable<PinnedBlocks>,
 ) =>
-  concat(
-    (_at
-      ? of(_at)
-      : chainHead.finalized$.pipe(
-          take(1),
-          map((x) => x.hash),
-        )
-    ).pipe(
-      mergeMap((at) =>
-        chainHead.validateTx$(at, tx).pipe(
-          map((isValid) => {
-            if (!isValid) throw new Error("Invalid")
-            return { type: "broadcasted" as "broadcasted" }
-          }),
-        ),
-      ),
-    ),
-    new Observable<{ type: "analyzed"; value: AnalyzedBlock }>((observer) => {
-      const subscription = chainHead
-        .trackTx$(tx)
-        .pipe(map((value) => ({ type: "analyzed" as const, value })))
-        .subscribe(observer)
-      subscription.add(
-        broadcastTx$(tx).subscribe({
-          error(e) {
-            observer.error(e)
-          },
-        }),
+  new Observable<
+    | {
+        hash: string
+        index: number
+        events: any
+      }
+    | boolean
+  >((observer) => {
+    const analyzedBlocks = new Map<string, AnalyzedBlock>()
+    let pinnedBlocks: PinnedBlocks
+    let latestState:
+      | {
+          hash: string
+          index: number
+          events: any
+        }
+      | boolean
+
+    const computeNextState = () => {
+      let current: string = pinnedBlocks.best
+      let analyzed: AnalyzedBlock | undefined = analyzedBlocks.get(current)
+
+      while (!analyzed) {
+        const block = pinnedBlocks.blocks.get(current)
+        if (!block) break
+        analyzed = analyzedBlocks.get((current = block.parent))
+      }
+
+      if (!analyzed) return // this shouldn't happen, though
+
+      const isFinalized =
+        pinnedBlocks.blocks.get(analyzed.hash)!.number <=
+        pinnedBlocks.blocks.get(pinnedBlocks.finalized)!.number
+
+      const found = analyzed.found.type
+      if (
+        found &&
+        typeof latestState === "object" &&
+        latestState.hash === analyzed.hash
+      ) {
+        if (isFinalized) observer.complete()
+        return
+      }
+
+      observer.next(
+        (latestState = found
+          ? {
+              hash: analyzed.hash,
+              ...analyzed.found,
+            }
+          : analyzed.found.isValid),
       )
-      return subscription
-    }),
-  )
+
+      if (isFinalized) {
+        if (found) observer.complete()
+        else if (!analyzed.found.isValid) observer.error(new Error("Invalid"))
+      }
+    }
+
+    const subscription = blocks$
+      .pipe(
+        distinctUntilChanged(
+          (a, b) => a.finalized === b.finalized && a.best === b.best,
+        ),
+      )
+      .subscribe({
+        next: (pinned: PinnedBlocks) => {
+          pinnedBlocks = pinned
+          if (analyzedBlocks.size === 0) return
+          computeNextState()
+        },
+        error(e) {
+          observer.error(e)
+        },
+      })
+
+    subscription.add(
+      analized$.subscribe({
+        next: (block) => {
+          analyzedBlocks.set(block.hash, block)
+          computeNextState()
+        },
+        error(e) {
+          observer.error(e)
+        },
+      }),
+    )
+
+    return subscription
+  }).pipe(distinctUntilChanged((a, b) => a === b))
 
 const getTxSuccessFromSystemEvents = (
   systemEvents: Array<SystemEvent>,
@@ -74,62 +137,92 @@ const getTxSuccessFromSystemEvents = (
 export const submit$ = (
   chainHead: ChainHead$,
   broadcastTx$: (tx: string) => Observable<never>,
-  transaction: HexString,
+  tx: HexString,
   at?: HexString,
-): Observable<TxBroadcastEvent> =>
-  tx$(chainHead, broadcastTx$, transaction, at).pipe(
-    mergeMap((result) => {
-      if (result.type === "broadcasted") return of(result)
+  emitSign = false,
+): Observable<TxEvent> => {
+  const txHash = hashFromTx(tx)
+  const getTxEvent = <
+    Type extends TxEvent["type"],
+    Rest extends Omit<TxEvent & { type: Type }, "type" | "txHash">,
+  >(
+    type: Type,
+    rest: Rest,
+  ): TxEvent & { type: Type } =>
+    ({
+      type,
+      txHash,
+      ...rest,
+    }) as any
 
-      if (!result.value.found.type) {
-        if (result.value.found.isValid) return EMPTY
-
-        return chainHead.isBestOrFinalizedBlock(result.value.hash).pipe(
-          filter((x) => x === "finalized"),
-          map(() => {
-            throw new Error("Invalid")
-          }),
-        )
-      }
-
-      const { index } = result.value.found
-      return combineLatest([
-        chainHead
-          .isBestOrFinalizedBlock(result.value.hash)
-          .pipe(filter(Boolean)),
-        result.value.found.events as Observable<Array<SystemEvent>>,
-      ]).pipe(
-        map(([type, events]) => ({
-          type: type === "best" ? ("bestChainBlockIncluded" as const) : type,
-          block: {
-            hash: result.value.hash,
-            index,
-          },
-          ...getTxSuccessFromSystemEvents(events, index),
-        })),
-      )
-    }),
-    takeWhile((e) => e.type !== "finalized", true),
+  const at$ = chainHead.pinnedBlocks$.pipe(
+    take(1),
+    map((blocks) => blocks.blocks.get(at!)?.hash ?? blocks.finalized),
   )
+
+  const validate$: Observable<never> = at$.pipe(
+    mergeMap((at) =>
+      chainHead.validateTx$(at, tx).pipe(
+        filter((x) => !x),
+        map(() => {
+          throw new Error("Invalid")
+        }),
+      ),
+    ),
+  )
+
+  const track$ = new Observable<AnalyzedBlock>((observer) => {
+    const subscription = chainHead.trackTx$(tx).subscribe(observer)
+    subscription.add(
+      broadcastTx$(tx).subscribe({
+        error(e) {
+          observer.error(e)
+        },
+      }),
+    )
+    return subscription
+  })
+
+  const bestBlockState$ = computeState(track$, chainHead.pinnedBlocks$).pipe(
+    map((x) => {
+      if (x === true || x === false)
+        return getTxEvent("txBestBlocksState", {
+          found: false,
+          isValid: x,
+        })
+
+      return getTxEvent("txBestBlocksState", {
+        found: true,
+        block: {
+          index: x.index,
+          hash: x.hash,
+        },
+        ...getTxSuccessFromSystemEvents(x.events, x.index),
+      })
+    }),
+  )
+
+  return concat(
+    emitSign ? of(getTxEvent("signed", {})) : EMPTY,
+    validate$,
+    of(getTxEvent("broadcasted", {})),
+    bestBlockState$.pipe(
+      continueWith(({ found, type, ...rest }) =>
+        found ? of(getTxEvent("finalized", rest as any)) : EMPTY,
+      ),
+    ),
+  )
+}
 
 export const submit = async (
   chainHead: ChainHead$,
   broadcastTx$: (tx: string) => Observable<never>,
   transaction: HexString,
   at?: HexString,
-): Promise<{
-  ok: boolean
-  events: Array<SystemEvent["event"]>
-  block: { hash: string; index: number }
-}> =>
+): Promise<TxFinalizedPayload> =>
   lastValueFrom(submit$(chainHead, broadcastTx$, transaction, at)).then((x) => {
     if (x.type !== "finalized") throw null
-    const result: {
-      ok: boolean
-      events: Array<SystemEvent["event"]>
-      block: { hash: string; index: number }
-      type?: any
-    } = { ...x }
-    delete result.type
+    const result: TxFinalizedPayload = { ...x }
+    delete (result as any).type
     return result
   })